Added basic header to Event module
- Moved WindowProps to skelly defs;
- Added headers to event and submodules;
